refactor TIM版本处理器以使用新的VersionBeanUtil
将TIM版本数据处理从单独的TIMVersionBeanUtil迁移到通用的VersionBeanUtil。此更改通过消除冗余代码，简化了数据处理逻辑。更新MainActivity以使用新的util进行TIM版本处理。
