+++ app/src/main/java/com/xiaoniu/qqversionlist/util/VersionBeanUtil.kt
@@ -20,11 +20,52 @@ package com.xiaoniu.qqversionlist.util
 
 import com.google.gson.Gson
 import com.google.gson.JsonObject
+import com.xiaoniu.qqversionlist.QVTApplication.Companion.EARLIEST_QQNT_FRAMEWORK_QQ_VERSION_STABLE
+import com.xiaoniu.qqversionlist.data.QQVersionBean
 import com.xiaoniu.qqversionlist.data.TIMVersionBean
 import com.xiaoniu.qqversionlist.ui.MainActivity
 import com.xiaoniu.qqversionlist.util.StringUtil.toPrettyFormat
+import kotlinx.serialization.json.Json
+import org.apache.maven.artifact.versioning.DefaultArtifactVersion
 
-object TIMVersionBeanUtil {
+object VersionBeanUtil {
+    fun qqVersionBeanProcessor(thisActivity: MainActivity, responseData: String) {
+        val start = (responseData.indexOf("versions64\":[")) + 12
+        val end = (responseData.indexOf(";\n" + "      typeof"))
+        val totalJson = responseData.substring(start, end)
+        thisActivity.qqVersion = totalJson.split("},{").reversed().map {
+            val pstart = it.indexOf("{\"versions")
+            val pend = it.indexOf(",\"length")
+            val json = it.substring(pstart, pend)
+            Json.decodeFromString<QQVersionBean>(json).apply {
+                jsonString = json
+                // 标记本机 Android QQ 版本
+                this.displayInstall = (DataStoreUtil.getStringKV(
+                    "QQVersionInstall", ""
+                ) == this.versionNumber)
+                this.isAccessibility = false
+                // 无障碍标记
+                /*DefaultArtifactVersion(this.versionNumber) >= DefaultArtifactVersion(
+                    EARLIEST_ACCESSIBILITY_QQ_VERSION
+                )*/
+
+                this.isQQNTFramework =
+                    DefaultArtifactVersion(this.versionNumber) >= DefaultArtifactVersion(
+                        EARLIEST_QQNT_FRAMEWORK_QQ_VERSION_STABLE
+                    )
+            }
+        }
+        if (DataStoreUtil.getBooleanKV(
+                "displayFirst", true
+            )
+        ) thisActivity.qqVersion[0].displayType = 1
+        // 舍弃 currentQQVersion = qqVersion.first().versionNumber
+        // 大版本号也放持久化存储了，否则猜版 Shortcut 因为加载过快而获取不到东西
+        DataStoreUtil.putStringKVAsync(
+            "versionBig", thisActivity.qqVersion.first().versionNumber
+        )
+    }
+
     fun timVersionBeanProcessor(thisActivity: MainActivity, responseData: String) {
         val start = (responseData.indexOf("var params= ")) + 12
         val end = (responseData.indexOf(";\n" + "      typeof"))
@@ -123,10 +164,15 @@ object TIMVersionBeanUtil {
         }
 
         // 去除重复的版本号
-        thisActivity.timVersion = thisActivity.timVersion.distinctBy { it.jsonString.toPrettyFormat() }
+        thisActivity.timVersion =
+            thisActivity.timVersion.distinctBy { it.jsonString.toPrettyFormat() }
+        if (thisActivity.timVersion[0].version == thisActivity.timVersion[1].version && thisActivity.timVersion[0].fix == "") (thisActivity.timVersion as MutableList<TIMVersionBean>).removeAt(
+            0
+        )
 
         if (DataStoreUtil.getBooleanKV(
-                "displayFirst", true
+                "displayFirst",
+                true
             )
         ) thisActivity.timVersion[0].displayType = 1
         DataStoreUtil.putStringKVAsync(
